Implement MVC pattern
This commit includes:
reorganize files to follow mvc
change express-validator in favor of validator
better use mongoose schema
