+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "freegamesnewsletter-site",
-  "version": "2.0.1",
+  "version": "3.0.0",
   "description": "",
   "main": "src/server.js",
   "scripts": {
@@ -20,8 +20,8 @@
     "dotenv": "^8.2.0",
     "ejs": "^3.0.2",
     "express": "^4.17.1",
-    "express-validator": "^6.4.0",
     "helmet": "^3.23.1",
-    "mongoose": "^5.9.9"
+    "mongoose": "^5.9.9",
+    "validator": "^13.1.1"
   }
 }
