+++ src/models/broker.js
@@ -1,7 +1,7 @@
 var amqp = require('amqplib/callback_api');
 var amqpConn = null;
 
-const CONN_URL = process.env.RABBITMQ_URL;
+const CONN_URL = 'amqp://rabbitmq/';
 const RABBITMQ_USER = process.env.RABBITMQ_USER;
 const RABBITMQ_PASS = process.env.RABBITMQ_PASS;
 
@@ -10,21 +10,21 @@ function start() {
         credentials: amqp.credentials.plain(
             RABBITMQ_USER, RABBITMQ_PASS)
     };
-    amqp.connect(CONN_URL + "?heartbeat=60", opt, function(err, conn) {
+    amqp.connect(CONN_URL + '?heartbeat=60', opt, function(err, conn) {
         if (err) {
-            console.log("[AMQP]", err.message);
+            console.log('[AMQP]', err.message);
             return setTimeout(start, 1000);
         }
-        conn.on("error", function(err) {
-            if (err.message !== "Connection closing") {
-                console.log("[AMQP] conn error", err.message);
+        conn.on('error', function(err) {
+            if (err.message !== 'Connection closing') {
+                console.log('[AMQP] conn error', err.message);
             }
         });
-        conn.on("close", function() {
-            console.log("[AMQP] reconnecting");
+        conn.on('close', function() {
+            console.log('[AMQP] reconnecting');
             return setTimeout(start, 1000);
         });
-        console.log("[AMQP] connected");
+        console.log('[AMQP] connected');
         amqpConn = conn;
         startPublisher();
     });
@@ -37,11 +37,11 @@ function startPublisher() {
             return console.log(err);
         }
         
-        ch.on("error", function(err) {
-            console.log("[AMQP] channel error", err.message);
+        ch.on('error', function(err) {
+            console.log('[AMQP] channel error', err.message);
         });
-        ch.on("close", function() {
-            console.log("[AMQP] channel closed");
+        ch.on('close', function() {
+            console.log('[AMQP] channel closed');
         });
 
         pubChannel = ch;
@@ -53,11 +53,13 @@ function publish(exchange, routingKey, content) {
     pubChannel.publish(exchange, routingKey, msg, { persistent: true }, 
         function(err, ok) {
             if (err) {
-                return console.log("[AMQP] publish", err);
+                return console.log('[AMQP] publish', err);
             }
 
-            console.log("[AMQP] publish", ok);
+            console.log('[AMQP] publish', ok);
         });
 }
 
-module.exports = {start, publish};
+start();
+
+module.exports = {publish};
