+++ src/utils/hashUtils.js
